+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ Unreleased
 
 **Added**
 
+* Added ``Header`` and ``Time`` data types.
+* Added ROS API method to retrieve current ROS time: ``ros.get_time``.
+
 **Fixed**
 
 1.1.0